Component refactor and modularization
